Safari-609.1.11.1/LayoutTests/css3/flexbox/multiline-min-preferred-width.html

<!DOCTYPE html>
<html>
<head>
<link rel="stylesheet" href="resources/flexbox.css">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<style>
.item {
    padding: 7px;
    background-color: green;
    border: 4px solid olive;
    width: min-content;
}

.square {
    background-color: lime;
    width: 20px;
    height: 20px;
}
</style>
</head>
<body onload="checkLayout('.container')">
<div id=log></div>

<div class="container flexbox wrap" data-expected-height=42>
  <div class="item flexbox wrap" data-expected-width=42 data-expected-height=42>
    <div class="square" data-expected-width=20 data-expected-height=20></div>
  </div>
</div>

</body>
</html>
